Made sale object class its own file
